Add files via upload
Begins by asking the user if they want to start with or without a game--these options are practically the same right now. Later, the "Start with a Game" option might be used for more specific tracking and such. Not sure. This initial window also allows the user to quit the program.
After beginning the program, it should wait 10 seconds and then prompt the user if they want to take a break--if they say "yes", initiate a 5 second break. If not, track it. Again, after 10 seconds, the user will be prompted if they want to take a break, and this general process SHOULD repeat indefinitely (but it doesn't in testing for unknown reasons). After 3 "No" response are given, a 10 second break is forced, but It looks like the "No" button doesn't work. Cool.
A break consists of an almost inescapable screen telling the user to take a break. It can be exited by using alt tab. 
There is also a constant option to quit, besides when in a break.

Anyways, I'm gonna focus on the "limiting rage moments" part of the project now.